Changing Grind quests
Now Grind quests will work better. Instead of ending as soon as the listener fires, it now fires a Complete script which flags the player to show that they have killed all necessary entities. It can also receive a new definition, endOnComplete, which will allow it to assume the old functionality of ending as soon as the player kills their last entity.